requirements.txt
Pillow==6.1.0
pyperclip==1.7.0
